requirements.txt
# AI-Vaani — Python dependencies
# Install with: pip install -r requirements.txt

# Core framework
streamlit>=1.35.0

# Environment variable loading
python-dotenv>=1.0.0

# LLM inference (Groq — free tier, 500+ tok/sec)
groq>=0.9.0

# Audio capture from browser microphone
streamlit-webrtc>=0.47.6
av>=11.0.0

# HTTP client for Speechmatics API
requests>=2.31.0

# Numerical audio processing
numpy>=1.26.0

# Optional: Google Cloud STT alternative
# google-cloud-speech>=2.26.0

# Optional: local audio file support
# soundfile>=0.12.1
# librosa>=0.10.2
